Unstyled CMS pages
This CMS uses MySQL database entries to identify users by id and corresponding image files and locations stored in a file system. The purpose of this system is to be able to create user accounts for family members and have them upload pictures which will be displayed on personal web pages.